Deprecate scrapy.utils.gz.
