+++ scrapy/utils/_compression.py
@@ -1,6 +1,11 @@
+from __future__ import annotations
+
+import struct
 import sys
 import zlib
+from gzip import GzipFile
 from io import BytesIO
+from typing import TYPE_CHECKING
 
 try:
     import brotli
@@ -11,6 +16,9 @@ if sys.version_info >= (3, 14):
     from compression import zstd
 else:
     from backports import zstd
+
+if TYPE_CHECKING:
+    from scrapy.http import Response
 
 
 _CHUNK_SIZE = 65536  # 64 KiB
@@ -32,6 +40,35 @@ class _DecompressionMaxSizeExceeded(ValueError):
 def _check_max_size(decompressed_size: int, max_size: int) -> None:
     if max_size and decompressed_size > max_size:
         raise _DecompressionMaxSizeExceeded(decompressed_size, max_size)
+
+
+def gunzip(data: bytes, *, max_size: int = 0) -> bytes:
+    """Gunzip the given data and return as much data as possible.
+
+    This is resilient to CRC checksum errors.
+    """
+    f = GzipFile(fileobj=BytesIO(data))
+    output_stream = BytesIO()
+    chunk = b"."
+    decompressed_size = 0
+    while chunk:
+        try:
+            chunk = f.read1(_CHUNK_SIZE)
+        except (OSError, EOFError, struct.error):
+            # complete only if there is some data, otherwise re-raise
+            # see issue 87 about catching struct.error
+            # some pages are quite small so output_stream is empty
+            if output_stream.getbuffer().nbytes > 0:
+                break
+            raise
+        decompressed_size += len(chunk)
+        _check_max_size(decompressed_size, max_size)
+        output_stream.write(chunk)
+    return output_stream.getvalue()
+
+
+def gzip_magic_number(response: Response) -> bool:
+    return response.body[:3] == b"\x1f\x8b\x08"
 
 
 def _inflate(data: bytes, *, max_size: int = 0) -> bytes:
